change log file location
