+++ packages/debug-test/Utilities.ts
@@ -30,3 +30,18 @@ export function abbreviate(s: string, maxLength: number): string {
   }
   return s.slice(0, maxLength - /* length of "..." */ 3) + '...';
 }
+
+/**
+ * Formats the given date object as a "YYYYMMDD-hhmmss" string.
+ * @param dt Date to format.
+ * @returns String representation of the data.
+ */
+export function formatDateTime(dt: Date) {
+  const YYYY = ('000' + dt.getFullYear()).slice(-4);
+  const MM = ('0' + (dt.getMonth() + 1)).slice(-2);
+  const DD = ('0' + dt.getDate()).slice(-2);
+  const hh = ('0' + dt.getHours()).slice(-2);
+  const mm = ('0' + dt.getMinutes()).slice(-2);
+  const ss = ('0' + dt.getSeconds()).slice(-2);
+  return `${YYYY}${MM}${DD}-${hh}${mm}${ss}`;
+}
